审批 json  异常  服务端 json数据对不上的问题

+++ app/src/main/java/com/loyo/oa/v2/beans/WfInstance.java
@@ -57,10 +57,10 @@ public class WfInstance extends BaseBeans implements Serializable {
     }
 
     public class Demand {
-        public int actualNum;
-        public int actualPrice;
-        public int estimatedNum;
-        public int estimatedPrice;
+        public float actualNum;
+        public float actualPrice;
+        public float estimatedNum;
+        public float estimatedPrice;
         public String memo;
         public String unit;
         public String productName;
